Refactor: extract manifest and deduplicate campaign validation
- Extract Manifest() to manifest.go (315 lines) — sendgrid.go drops
  from 445 to 137 lines, making it easier to find connector setup code
- Extract shared campaignFields struct and validate() to campaign.go —
  eliminates duplicated validation between send_campaign and
  schedule_campaign (was ~20 identical lines in each)
- schedule_campaign now embeds campaignFields and delegates shared
  validation, only adding send_at-specific checks
- buildSingleSendBody now takes *campaignFields instead of a
  sendCampaignParams, making it reusable from both actions
